backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Pydantic
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-auth==2.38.0

# Kafka
aiokafka==0.12.0

# Redis & Celery
redis==5.2.1
celery==5.4.0

# AI
google-generativeai==0.8.4
httpx==0.28.1

# AWS
aioboto3==13.4.0

# Rate limiting
slowapi==0.1.9

# Utils
loguru==0.7.3
python-dotenv==1.0.1
firebase-admin==6.6.0
